requirements.txt
Django==2.1.4
djangorestframework==3.9.0
django-split-settings==0.3.0
psycopg2-binary==2.7.6.1
django-cors-headers==2.4.0
drf-yasg==1.12.1
